chore: update package dependencies and refactor server implementation
- Added `ts-unused-exports` package to manage unused TypeScript exports.
- Removed the `custom-server.ts` file and integrated its functionality into `server.ts`.
- Refactored `imdbMetaProvider` and `cinemetaMetaProvider` functions to improve code clarity.
- Updated `package.json` and `package-lock.json` to reflect new dependencies.
