+++ package.json
@@ -36,6 +36,7 @@
     "@types/node": "^20.14.10",
     "beamup-cli": "^1.3.0",
     "prettier": "^3.3.2",
+    "ts-unused-exports": "^11.0.1",
     "tsx": "^4.16.2",
     "typescript": "^5.8.3",
     "vitest": "^2.1.5"
